refactor(dashboard-customize): separate it from page to reuse

refactor(dashboard-customize): separate it from page to reuse
